[TASK] Adjust coding styles (#18)
These are now the same as in the render-guides project. We stick to PER-CS1 with the already available implemented changes for PER-CS2 - so there are no changes to the code for now. This prevents necessary code style adjustments when more rules are added for PER-CS2 in future releases of php-cs-fixer. We can opt-in to them explicitely if desired and switch later to PER-CS2 when the ruleset is completely available in php-cs-fixer